requirements.txt
requests
pandas
numpy
requests
python-dotenv
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
pycryptodome